Already added error reports file:line
When Add, AddErr, New or NewErr or its callers fail because
the added injection type is already in the graph, the
error/panic includes the file and line where that type was
previously added.

Unfortunately this introduces brittle tests that depend on file
names and line numbers. It may be possible to address these in
another way, but it's essential to test that these are reported
correctly.